Feature/security

+++ Blog/src/main/java/com/noseryoung/uek223/domain/blogPost/BlogPostServiceImpl.java
@@ -1,7 +1,14 @@
 package com.noseryoung.uek223.domain.blogPost;
 
+import com.noseryoung.uek223.domain.appUser.User;
+import com.noseryoung.uek223.domain.appUser.UserRepository;
+import com.noseryoung.uek223.domain.appUser.UserService;
 import com.noseryoung.uek223.domain.blogPost.dto.UpdateBlogPostDTO;
+import com.noseryoung.uek223.domain.exceptions.InvalidObjectException;
+import com.noseryoung.uek223.domain.exceptions.NoAccessException;
 import com.noseryoung.uek223.domain.exceptions.NoBlogPostFoundException;
+import com.noseryoung.uek223.domain.role.Role;
+import com.noseryoung.uek223.domain.role.RoleRepository;
 import com.noseryoung.uek223.domain.utils.LevenshteinDistance;
 import com.noseryoung.uek223.domain.utils.LevenshteinResult;
 import com.noseryoung.uek223.domain.utils.MultiStopwatch;
@@ -13,8 +20,12 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.util.*;
 
 @Service
@@ -23,23 +34,32 @@ public class BlogPostServiceImpl implements BlogPostService {
 
     private final BlogPostRepository blogPostRepository;
     private final BlogPostMapper blogPostMapper;
+    private final UserRepository userRepository;
     private final NullAwareBeanUtilsBean nullAwareBeanUtilsBean;
 
     @Override
-    public BlogPost create(BlogPost blogPost) {
+    public BlogPost createBlogPost(BlogPost blogPost) {
         return blogPostRepository.saveAndFlush(blogPost);
     }
 
     @Override
-    @SneakyThrows
-    public BlogPost update(UpdateBlogPostDTO blogPost, UUID id) {
-
-        BlogPost newBlogPost = blogPostMapper.updateBlogPostDTOToBlog(blogPost);
-        BlogPost oldBlogPost = blogPostRepository.findById(id).orElseThrow(() -> new NoBlogPostFoundException("No BlogPost found with the given id"));
-
-        nullAwareBeanUtilsBean.copyProperties(oldBlogPost, newBlogPost);
-
-        return blogPostRepository.saveAndFlush(oldBlogPost);
+    public BlogPost updateBlogPost(UpdateBlogPostDTO blogPost, UUID id) throws NoAccessException, NoBlogPostFoundException, InvalidObjectException {
+        if (blogPostRepository.existsById(id)){
+            if (hasAccess(id)){
+                BlogPost newBlogPost = blogPostMapper.updateBlogPostDTOToBlog(blogPost);
+                BlogPost oldBlogPost = findById(id);
+                try {
+                    nullAwareBeanUtilsBean.copyProperties(oldBlogPost, newBlogPost);
+                } catch (Exception e){
+                    throw new InvalidObjectException("An unexpected error occurred. Please verify that the provided blogpost is valid!");
+                }
+                return blogPostRepository.saveAndFlush(oldBlogPost);
+            } else {
+                throw new NoAccessException();
+            }
+        } else {
+            throw new NoBlogPostFoundException();
+        }
     }
 
     @Override
@@ -89,14 +109,30 @@ public class BlogPostServiceImpl implements BlogPostService {
     }
 
     @Override
-    public void delete(UUID id) {
-        blogPostRepository.deleteById(id);
+    public void deleteBlogPost(UUID id) throws NoAccessException, NoBlogPostFoundException {
+        if (blogPostRepository.existsById(id)) {
+            if (hasAccess(id)) {
+                blogPostRepository.deleteById(id);
+            } else {
+                throw new NoAccessException();
+            }
+        } else {
+            throw new NoBlogPostFoundException();
+        }
     }
 
     @Override
     public List<BlogPost> findAll(int page, int length) {
+        //TODO: change sort by title to sort by publish date
         Pageable pageable = PageRequest.of(page, length, Sort.by("title").ascending());
         return blogPostRepository.findAll(pageable).getContent();
     }
+
+    private boolean hasAccess(UUID id) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        return findById(id).getAuthor().getId().equals(userRepository.findByUsername(auth.getName()).getId()) ||
+                auth.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ADMIN"));
+    }
+
 
 }
